Small graphical improvements
I have been experimenting with some graphics and have started building a
custom theme.

+++ Todo2/Code3/src/com/todo/code3/misc/App.java
@@ -53,7 +53,7 @@ public class App {
 	public static final int OPTIONS_MOVE = 3;
 
 	// converting dp to pixels and vice versa
-	public static int dpToPx(int dp, Resources r) {
+	public static int dpToPx(float dp, Resources r) {
 		DisplayMetrics displayMetrics = r.getDisplayMetrics();
 		int px = Math.round(dp * (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT));
 		return px;
